refactor: reorganize Security config classes
