+++ src/main/java/eu/bbmri_eric/negotiator/common/configuration/security/auth/MvcConfig.java
@@ -1,0 +1,14 @@
+package eu.bbmri_eric.negotiator.common.configuration.security.auth;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.web.servlet.util.matcher.MvcRequestMatcher;
+import org.springframework.web.servlet.handler.HandlerMappingIntrospector;
+
+@Configuration
+public class MvcConfig {
+  @Bean
+  MvcRequestMatcher.Builder mvc(HandlerMappingIntrospector introspector) {
+    return new MvcRequestMatcher.Builder(introspector);
+  }
+}
